common: add Broadcaster and OrchestratorPool interfaces to common types
